+++ addOns/oast/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.0] - 2022-01-10
 ### Changed
 - Set HttpSender's initiator to `OAST_INITIATOR`, value 16.
 
@@ -83,6 +83,7 @@ Button is disabled in that case.
 
 ## [0.1.0] - 2021-08-04
 
+[0.8.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.8.0
 [0.7.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.7.0
 [0.6.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.6.0
 [0.5.0]: https://github.com/zaproxy/zap-extensions/releases/oast-v0.5.0